+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 When making changes, add an entry to the **Unreleased** section.
 
+## [Unreleased]
+
+### Changed
+
+- Grader main script now unzips to temporary directory instead of to the shared volume
+
 ## [0.8.2] - 2025-07-10
 
 ### Changed